% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqldump.R
\name{dumpSQL}
\alias{dumpSQL}
\title{Import session-based SQL files in a MySQL database}
\usage{
dumpSQL(path_to_data, type = c("general", "forum", "unanonymizable",
  "hash_mapping"), user = "root", password = "root", verbose = FALSE)
}
\arguments{
\item{path_to_data}{Folder with data files in it}

\item{type}{Which dataset should be stored?}

\item{user}{Username for MySQL database. Defaults to 'root'}

\item{password}{Password for MySQL database. Defaults to 'root'}

\item{verbose}{Print verbose intermediate messages? Defaults to FALSE}
}
\description{
This function imports session-based SQL files into a MySQL database.
}
\details{
This function requires a running MySQL database (see: https://www.mysql.com/)
}
\examples{
\dontrun{
# Dump forum SQL file for federalism course
dumpSQL("/home/jasper/Documents/MOOC-DATA/Data_Dumps/federalism-001","forum",verbose = TRUE)
}
}
\author{
Jasper Ginn
}
\seealso{
\code{\link{mongodump}}
}

