% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mongodump.R
\name{mongodump}
\alias{mongodump}
\title{Import clickstream data into a mongodb}
\usage{
mongodump(path_to_clickstream, database = "moocs", create_index = TRUE,
  index_value = "username", verbose = FALSE, jsonArray = FALSE, ...)
}
\arguments{
\item{path_to_clickstream}{Path to compressed archive with clickstream file}

\item{database}{database in which to store the collection of clickstream data (collection will be defined as course name)}

\item{create_index}{Mongo can create an index on usernames, which speeds up queries considerably. Defaults to TRUE}

\item{index_value}{Which value will mongodb use for indexing?}

\item{verbose}{Print verbose intermediate messages? Defaults to FALSE}

\item{jsonArray}{Use the --jsonArray flag while importing documents to mongo. This treats the input as a jsonArray (see https://docs.mongodb.org/manual/reference/program/mongoimport/)}

\item{...}{any other arguments needed to connect to mongodb (e.g. username, password, host)}
}
\description{
This function imports session-based JSON clickstream files into a mongodb database.
}
\details{
This function requires a running mongodb database (see: https://www.mongodb.org/)
}
\examples{
\dontrun{
# Dump clickstream file into mongo
mongodump("/home/jasper/Documents/MOOC-DATA/Data_Dumps/configuringworld-001/configuringworld-001_clickstream_export.gz",
verbose = TRUE)
}
}
\author{
Jasper Ginn
}
\seealso{
\code{\link{sqldump}}
}

